Add event listeners to number buttons
Query selected all elements of the calculator that the user can interact
with and assigned them to an appropriate constant. Then added event
listeners to each of the number buttons so that when the button is
clicked by the user, the corresponding number is appended to the users
current input string. This string can then be used for calculations and
for display on the calculators screen.

Updated various class and id values in index.html for consistency.
Commented out button border styling in style.css temporarily so that
button clicks are more visible whilst testing.