Major calculation bug fix
Encountered a major bug where the mathematical operation buttons would
cease to function. The bug would occur when:
- Input any calculation and hit = to get the answer
- Without pressing AC or any mathematical operation button, press a
number button to start a new number input.
- The mathematical operations would now cease to function.

The cause was how the equals and mathematical functions were coded.
Equals was coded to set the output to previousInput and then clear both
currentInput and the stored operator variable. When you then entered a
new number, it would be stored as currentInput. If you had a populated
currentInput and previousInput, then the mathematical functions were
also expecting to see operator populated. As it was blank, the functions
would fail to execute.
Fixed the bug by adding an additional code to the number buttons. If the
previousInput is populated and both operator and currentInput aren't,
then the number buttons will now run the reset() function prior to
inputting their respective numbers. This is likely the outcome a user
would anticipate as they likely intended to start a new calculation by
inputting a new number directly after calculating the output of the
previous calculation.

Additional change: updated the content sentence in index.html